moved VaultErrors and VaultWarnings to vault components
