+++ features/openMultiplyVault/variants/guni/open/GuniOpenMultiplyVaultForm.tsx
@@ -1,18 +1,18 @@
 import { VaultAllowance, VaultAllowanceStatus } from 'components/vault/VaultAllowance'
+import { VaultChangesWithADelayCard } from 'components/vault/VaultChangesWithADelayCard'
+import { VaultErrors } from 'components/vault/VaultErrors'
 import { VaultFormContainer } from 'components/vault/VaultFormContainer'
+import { VaultProxyStatusCard } from 'components/vault/VaultProxy'
+import { VaultWarnings } from 'components/vault/VaultWarnings'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
 
-import { VaultChangesWithADelayCard } from '../../../../../components/vault/VaultChangesWithADelayCard'
-import { VaultProxyStatusCard } from '../../../../../components/vault/VaultProxy'
 import { OpenMultiplyVaultButton } from '../../../common/OpenMultiplyVaultButton'
 import {
   OpenMultiplyVaultConfirmation,
   OpenMultiplyVaultStatus,
 } from '../../../common/OpenMultiplyVaultConfirmation'
 import { OpenMultiplyVaultTitle } from '../../../common/OpenMultiplyVaultTitle'
-import { VaultErrors } from '../../../common/VaultErrors'
-import { VaultWarnings } from '../../../common/VaultWarnings'
 import { OpenMultiplyVaultState } from '../../../openMultiplyVault'
 import { GuniOpenMultiplyVaultEditing } from './GuniOpenMultiplyVaultEditing'
